feat(nextly): persist plugin options declared on code-defined user fields
